requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
Markdown==3.4.3
numpy==1.24.2
opencv-python==4.7.0.68
Pillow==9.4.0
pypng==0.20220715.0
pytz==2023.2
qrcode==7.4.2
sqlparse==0.4.3
timeago==1.0.16
typing_extensions==4.5.0
